Forgot to typecast Math_PI

+++ core/math/math_funcs.h
@@ -446,7 +446,7 @@ public:
 	}
 	static _ALWAYS_INLINE_ float move_toward_angle(float p_from, float p_to, float p_delta) {
 		float difference = angle_difference(p_from, p_to);
-		return wrapf(p_from + fmin(abs(difference), p_delta) * SIGN(difference), -Math_PI, Math_PI);
+		return wrapf(p_from + fmin(abs(difference), p_delta) * SIGN(difference), -(float)Math_PI, (float)Math_PI);
 	}
 
 	static _ALWAYS_INLINE_ double linear_to_db(double p_linear) {
